Code _CNN/requirements.txt
matplotlib == 3.5.2
numpy == 1.23.1
opencv-python == 4.6.0.66
Pillow == 9.4.0
scipy == 1.9.0
torch == 1.13.0
torchvision == 0.14.0
wandb == 0.13.2
seaborn == 0.11.2
pandas == 1.4.3
PyYAML == 6.0
tqdm == 4.64.0
